Fix make dist check
Many files are not listed as required in the tarball distribution,
leading to failures in 'make distcheck'.  This patch adds them to the
relevant '_SOURCES' variables so that they get distributed in the
tarball, but not installed with 'make install' (see libtool docs [1]).
Additionally, this patch removes the inclusion of 'arith128.h' from
'src/tipowof10.c', because it is not needed.

Signed-off-by: Gabriel F. T. Gomes <gabriel@inconstante.net.br>

[1] https://www.gnu.org/software/automake/manual/automake.html#Libtool-Libraries