Add missing files to the distribution
New files have been added to pveclib, but not to EXTRA_DIST fields,
which cause 'make distcheck' to fail. This patch adds them to
EXTRA_DIST, accordingly, as well as it adds a new make target,
distclean-local, to clean files built with regular make rules, rather
than with libtool's macros.

Signed-off-by: Gabriel F. T. Gomes <gabriel@debian.org>