Update-p2 Float128 divide round to odd improvement. Close #196
The Power8/9 implementation of divide extended quadword
(vec_vdiveuq_inline() was improved. This was the basis for the special
round to odd version (vec_diveuq_qpo()) used in float128
vec_xsdivqpo_inline()) This is sub-optimal for this case.

In the extended divide case, and the divisor has two nonzero
(doubleword) digits, an explicit multiple/subtract is required to
obtain the remainder. The optimized implementation needs a
(128 by 64-bit) multiply that is not defined within vec-int128_ppc.h
but can be constructed using vec_vmuloud() and vec_vmaddeud().
The vec_diveuq_qpo() implementation uses this method and also checks
the final remainer for nonzero. A nonzero remainder causes the LSB 1
to be ORed into the quotient (Round-to-Odd).

Signed-off-by: Steven Munroe <munroesj52@gmail.com>