Add exploit common parts for __int128 and P10. After review.
The DP/QP floating-point codes need a number of special values and
masks that difficult to generated. These are often of the form
(2**n -1) and can be generated with -1 const and shift right.
This update adds special cases for 31, 63, 127, 255 to
vec_splat_u128() and vec_splat_u64. Larger const masks will be
handled in vec_f64_ppc.h and vec_f128_ppc.h updates as they are
specific to FP exponents and bias values.

	* src/pveclib/vec_int128_ppc.h (vec_splat_u128): Add special
	cases for 31, 63, 127, 255.

	* src/pveclib/vec_int64_ppc.h (vec_splat_u64[_ARCH_PWR8]):
	Add special cases for 31, 63, 127, 255.

	* src/testsuite/vec_int128_dummy.c (__test_splatiuq_31_V1,
	__test_splatiuq_63_V1, __test_splatiuq_127_V1): New compile test.

	* src/testsuite/vec_int64_dummy.c (__test_splatiud_31,
	__test_splatiud_31_V1, __test_splatiud_52, __test_splatiud_63,
	__test_splatiud_63_V1, __test_splatiud_127_V1): New compile test.

	* src/testsuite/vec_pwr9_dummy.c (__test_splatiuq_PWR9_V5,
	__test_splatiud_PWR9_V4): New compile test.

Signed-off-by: Steven Munroe <munroesj52@gmail.com>