Continue updates to vec_float128, add compiler optimizations.
The xfer and mask operations between __float128 and vector int using
unions exposed poor optimizations starting with GCC 8 and continuing
through GCC 11. For the POWER7/8 target GCC would generate vector
store then reload sequences for what should be a simple vector
register transfer. This had devastating impacts on performance.

This appears to be fixed for GCC 12 and 13 and I assume 14/15.
So this update changes the macros logic to reenable the union transfers
for those versions of __GNUC__.

	* src/pveclib/vec_f128_ppc.h (vec_sel_bin128_2_bin128,
	vec_and_bin128_2_vui32t, vec_andc_bin128_2_vui32t,
	vec_or_bin128_2_vui32t, vec_xor_bin128_2_vui32t,
	vec_andc_bin128_2_vui128t, vec_xfer_bin128_2_vui8t,
	vec_xfer_bin128_2_vui32t, vec_mrgh_bin128_2_vui64t,
	vec_mrgl_bin128_2_vui64t, vec_xfer_bin128_2_vui64t,
	vec_xfer_bin128_2_vui128t [__GNUC__]):
	Restrict generation of inline_asm for 7 > __GNUC__ < 12.
	Otherwise use union __VF_128.
	(vec_absf128[!_ARCH_PWR9]): Update power8 latency.
	Replace vec_mask128_f128sign with vec_mask128_f128mag and
	vec_and_bin128_2_vui32t.

Signed-off-by: Steven Munroe <munroesj52@gmail.com>